requirements.txt
twarc==2.12.0
click==8.1.3
gdown==4.4.0
pandas==1.5.2
nltk==3.7
numpy==1.23.5
scikit-learn==1.1.3
scipy==1.9.3
pyspellchecker==0.7.0
pyLDAvis==3.3.1
tensorflow==2.11.0 ; sys_platform=="linux"
tensorflow-macos==2.11.0 ; sys_platform=="darwin"
sentencepiece
plotly==5.11.0
pyarrow==10.0.1
geopy==2.3.0
geopandas==0.12.1
stanza==1.4.2
labMTsimple==2.8.7
matplotlib==3.6.2
emoji==2.2.0
rtree==1.0.1
mapclassify==2.4.3
spacy==3.4.3
happytransformer==2.4.1
timezonefinder==6.1.3
tweet-preprocessor==0.6.0
python-dotenv==0.21.0
flair==0.11.3
pandarallel==1.6.3
